[Fix] #525 - transform 코드 Network 모듈에서 Data 모듈로 변경
