+++ SOPT-iOS/Projects/Modules/Networks/Sources/Entity/CoreAuth/SocialAccountResultEntity.swift
@@ -7,15 +7,7 @@
 //
 
 import Foundation
-import Domain
 
 public struct SocialAccountResultEntity: Decodable {
-    let platform: PlatformType
+    public let platform: PlatformType
 }
-
-extension SocialAccountResultEntity {
-    public func toDomain()-> OAuthProvider {
-        self.platform.toDomain()
-    }
-}
-
